Added option to disable tee
Piping and tee may break some applications. Thus, it may be reasonable to
disable this feature in specific build systems. The optional boolean key "tee"
can be used to do this in the build system definition.

In the Readme an example for a simple terminal-based build system was added,
which can mimic the function of this package with `"tee": false`.